% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/httping-package.R
\docType{package}
\name{httping-package}
\alias{httping}
\alias{httping-package}
\title{Ping urls to time requests}
\description{
Ping urls to time requests
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

