% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{ping}
\title{Ping a url, doing a single call, with any http verbs}
\usage{
ping(url, verb = GET, ...)
}
\arguments{
\item{url}{A url}

\item{verb}{An http verb, default: \code{\link[httr]{GET}}}

\item{...}{Any httr verb parameters passed on to those functions}
}
\description{
Ping a url, doing a single call, with any http verbs
}
\details{
not sure this function is worth having, doesn't do a whole lot...
}
\examples{
\donttest{
"http://localhost:9200" \%>\% ping()
"http://localhost:9200" \%>\% ping(config=verbose())
"http://localhost:9200" \%>\% ping(config=c(verbose(), accept_json()))

"http://localhost:5984" \%>\% ping()
"http://localhost:5984" \%>\% ping(verb=HEAD)
"http://localhost:5984" \%>\% ping(verb=PUT)
"http://google.com" \%>\% ping()

# pass just a port number, tries to resolve, fails if not found
"9200" \%>\% ping()
9200 \%>\% ping()
9200 \%>\% ping(verb=POST)
9200 \%>\% ping(verb=HEAD)
5984 \%>\% ping()
ping(9200)
ping("9200")
}
}

