% Generated by roxygen2 (4.0.2): do not edit by hand
\name{time}
\alias{time}
\title{Ping a url to time the request}
\usage{
time(.request, count = 10, delay = 1, flood = FALSE, verbose = TRUE,
  ...)
}
\arguments{
\item{.request}{A httr response object}

\item{count}{integer, Number of requests to do.}

\item{delay}{integer, Seconds to delay successive calls by.}

\item{flood}{logical; If TRUE, no delay between requests. If FALSE, delay by 1 second.}

\item{verbose}{logical; If TRUE, print progress.}

\item{...}{Further args passed on to functions in \code{httr}}
}
\description{
Ping a url to time the request
}
\examples{
\donttest{
GET("http://localhost:9200") \%>\% time()
GET("http://localhost:5984") \%>\% time()
GET("http://google.com") \%>\% time()
}
}

