% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httping-package.R
\docType{package}
\name{httping-package}
\alias{httping}
\alias{httping-package}
\title{httping}
\description{
Ping urls to time requests
}
\section{Pipes}{

You can use this function with or without pipes. We export the pipe
\code{\%>\%} in this package so you don't have to load it separately.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sckott/httping}
  \item Report bugs at \url{https://github.com/sckott/httping/issues}
}

}
\author{
\strong{Maintainer}: Scott Chamberlain \email{myrmecocystus@gmail.com}

}
\keyword{internal}
