% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ping}
\alias{ping}
\title{Ping a url, doing a single call, with any http verbs}
\usage{
ping(url, verb = GET, ...)
}
\arguments{
\item{url}{A url}

\item{verb}{An http verb, default: \code{\link[httr]{GET}}}

\item{...}{Any httr verb parameters passed on to those functions}
}
\description{
Ping a url, doing a single call, with any http verbs
}
\details{
not sure this function is worth having, doesn't do a whole lot...
}
\examples{
\donttest{
"http://localhost:9200" \%>\% ping()
"http://localhost:5984" \%>\% ping()
"http://google.com" \%>\% ping()
}
}

