% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time}
\alias{time}
\title{Ping a url to time the request}
\usage{
time(.request, count = 10, delay = 0.5, flood = FALSE, verbose = TRUE,
  ...)
}
\arguments{
\item{.request}{A \pkg{httr} response object}

\item{count}{integer, Number of requests to do.}

\item{delay}{integer, Seconds to delay successive calls by. Default:
0.5 seconds.}

\item{flood}{logical; If \code{TRUE}, no delay between requests. If \code{FALSE},
delay by 0.5
second.}

\item{verbose}{logical; If \code{TRUE}, print progress.}

\item{...}{Further args passed on to functions in \pkg{httr}}
}
\description{
Ping a url to time the request
}
\examples{
\dontrun{
GET("https://mockbin.com/request") \%>\% time()
GET("https://api.github.com") \%>\% time()
GET("http://google.com") \%>\% time()
}
}
